fix: database version 14 fixed

+++ cmd/migrate/migrations/000014_alter_users_with_roles.down.sql
@@ -1,2 +1,2 @@
-ALTER TABLE IF EXISTS users
-DROP COLUMN role_id;
+ALTER TABLE
+  IF EXISTS users DROP COLUMN role_id;
